+++ linotp/tokens/hmactoken.py
@@ -38,7 +38,6 @@ from flask_babel import gettext as _
 from linotp.lib.apps import (
     NoOtpAuthTokenException,
     create_google_authenticator,
-    create_oathtoken_url,
 )
 from linotp.lib.auth.validate import check_otp, check_pin
 from linotp.lib.challenges import Challenges
@@ -713,8 +712,6 @@ class HmacTokenClass(TokenClass):
         response_detail.update(info)
         response_detail["serial"] = self.getSerial()
 
-        tok_type = self.type.lower()
-
         otpkey = None
         if "otpkey" in info:
             otpkey = info.get("otpkey")
@@ -732,36 +729,19 @@ class HmacTokenClass(TokenClass):
                 p["otpkey"] = otpkey
                 p["serial"] = self.getSerial()
                 # label
-                goo_url = create_google_authenticator(p, user=user)
+                enrollment_url = create_google_authenticator(p, user=user)
 
-                response_detail["googleurl"] = {
+                enrollment_url_detail = {
                     "order": "0",
                     "description": _("OTPAuth Url"),
-                    "value": goo_url,
-                    "img": create_img(goo_url, width=250),
+                    "value": enrollment_url,
+                    "img": create_img(enrollment_url, width=250),
                 }
+                # add enrollment_url_detail to both `enrollment_url` and soon to be deprecated `googleurl`
+                response_detail["enrollment_url"] = enrollment_url_detail
+                response_detail["googleurl"] = enrollment_url_detail
 
             except NoOtpAuthTokenException as exx:
                 log.warning(exx)
-
-            oath_support = getFromConfig("OATHTokenSupport", "False") == "True"
-            if oath_support:
-                if user is not None:
-                    try:
-                        oath_url = create_oathtoken_url(
-                            user.login,
-                            user.realm,
-                            otpkey,
-                            tok_type,
-                            serial=self.getSerial(),
-                        )
-                        response_detail["oathurl"] = {
-                            "order": "2",
-                            "description": _("URL for OATH token"),
-                            "value": oath_url,
-                            "img": create_img(oath_url, width=250),
-                        }
-                    except Exception as ex:
-                        log.info("failed to set oath or google url: %r", ex)
 
         return response_detail